+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/settings/Settings.java
@@ -65,6 +65,7 @@ public class Settings {
 
     public static final BooleanSetting ENABLE_DOWNLOAD = new BooleanSetting("enable_download", true);
     public static final BooleanSetting ENABLE_DIRECT_DOWNLOAD = new BooleanSetting("enable_direct_download", false);
+    public static final BooleanSetting INLINE_DOWNLOAD_BUTTON = new BooleanSetting("inline_download_button", false);
     public static final BooleanSetting DOWNLOAD_USERNAME_FOLDER = new BooleanSetting("download_username_folder", false);
     // Should be kept empty by default as its handled in `StorageUtils.java`
     public static final StringSetting CUSTOM_DOWNLOAD_PATH = new StringSetting("custom_download_path", "");