+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/settings/SettingsStatus.java
@@ -202,6 +202,8 @@ public class SettingsStatus {
     public static void downloadVoiceMessage() { downloadVoiceMessage = true; }
     public static boolean downloadWithExternalDownloader = false;
     public static void downloadWithExternalDownloader() { downloadWithExternalDownloader = true; }
+    public static boolean inlineDownloadButton = false;
+    public static void inlineDownloadButton() { inlineDownloadButton = true; }
     public static boolean downloadSection(){return (SettingsStatus.downloadMedia || SettingsStatus.downloadWithExternalDownloader);}
 
     public static boolean hideNavigationButtons = false;