+++ extensions/instagram/src/main/java/app/morphe/extension/instagram/settings/preference/ScreenBuilder.java
@@ -635,6 +635,16 @@ public class ScreenBuilder {
                 )
         );
 
+        if (SettingsStatus.inlineDownloadButton) {
+            addPreference(
+                    helper.switchPreference(
+                            str("piko_inline_download_button"),
+                            str("piko_inline_download_button_desc"),
+                            Settings.INLINE_DOWNLOAD_BUTTON
+                    )
+            );
+        }
+
         addPreference(
                 helper.switchPreference(
                         str("piko_download_username_folder"),