+++ patches/src/main/kotlin/app/crimera/patches/instagram/misc/download/DownloadMediaPatch.kt
@@ -12,6 +12,7 @@ import app.crimera.patches.instagram.entity.originalSoundDataIntf.originalSoundD
 import app.crimera.patches.instagram.entity.trackDataIntf.trackDataIntfEntity
 import app.crimera.patches.instagram.entity.videoData.videoDataEntity
 import app.crimera.patches.instagram.misc.directMessage.saveAllMessages.saveAllMessagesPatch
+import app.crimera.patches.instagram.misc.download.inlineDownloadButton.inlineDownloadButtonPatch
 import app.crimera.patches.instagram.misc.hookFlags.hookFlagsPatch
 import app.crimera.patches.instagram.misc.overflowMenuButton.posts.addOverflowMenuButtonAttributes
 import app.crimera.patches.instagram.misc.overflowMenuButton.posts.debugOverflowButton.debugOverflowMenuButtonPatch
@@ -47,12 +48,15 @@ val downloadMediaPatch =
             hookOverflowMenuButton,
             debugOverflowMenuButtonPatch,
             hookReelOverflowMenuButton,
+            inlineDownloadButtonPatch,
         )
         compatibleWith(COMPATIBILITY_INSTAGRAM)
 
         execute {
 
             addOverflowMenuButtonAttributes("PIKO_DOWNLOAD", "downloadOverflowButton")
+            addOverflowMenuButtonAttributes("PIKO_DOWNLOAD_CURRENT", "downloadCurrentOverflowButton")
+            addOverflowMenuButtonAttributes("PIKO_DOWNLOAD_ALL", "downloadAllOverflowButton")
 
             // DM media downloader.
             GetDirectThreadMediaSaverModuleNameFingerprint.apply {